refactor(client): modularize connection lifecycle
