+++ packages/client/src/index.ts
@@ -1,158 +1,3 @@
-import type {
-  ApplicationNotifications,
-  CollectionChange,
-  CollectionChangesPage,
-  CollectionDescription,
-  CollectionFileMetadata,
-  CollectionOperation,
-  CollectionTypeDocument,
-  CreateViewSourceInput,
-  DeleteViewSourceInput,
-  DeleteViewSourceResult,
-  EncryptedRelayOperationRequest,
-  EncryptedRelayOperationResponse,
-  ExecuteViewInput,
-  GrantEncryption,
-  GrantScope,
-  JsonObject,
-  MdbaseAppManifest,
-  MdbaseDiagnostic,
-  MdbaseOperationEnvelope,
-  MdbaseOperationRequest,
-  QueryRecord,
-  RecordDocument,
-  ReadViewSourceInput,
-  SavedViewExecution,
-  SavedViewList,
-  SavedViewSourceDocument,
-  SyncChangesPage,
-  SyncMutation,
-  SyncMutationReceipt,
-  SyncSession,
-  SyncSnapshotPage,
-  TypePackInstallResult,
-  TypePackProvision,
-  UpdateViewSourceInput
-} from "@mdbase/connect-protocol";
-import {
-  DEFAULT_LOOPBACK_PORT,
-  ENCRYPTED_RELAY_PROTOCOL_VERSION,
-  RELAY_ENCRYPTION_SUITE
-} from "@mdbase/connect-protocol";
-import {
-  decryptRelayResponse,
-  encryptRelayRequest,
-  IndexedDbGrantKeyStore,
-  MemoryGrantKeyStore,
-  RelayCryptoError,
-  signAuthorityRequest,
-  validateGrantEncryption,
-  type GrantKeyStore
-} from "./crypto.js";
-import { abortableDelay } from "./async.js";
-import {
-  authorizationReturnToFromError,
-  isAuthorizationCallbackUrl
-} from "./authorization-url.js";
-import { MdbaseCollectionClient } from "./collection-client.js";
-import {
-  MdbaseConnectError,
-  unwrapOperation
-} from "./errors.js";
-import type {
-  ChangesInput,
-  CreateInput,
-  CreateTypeInput,
-  DeleteInput,
-  DeletePreflightResult,
-  DeleteProgressOptions,
-  DeleteResult,
-  MdbaseCollectionTransport,
-  MdbaseDesiredTimer,
-  MdbaseTimer,
-  MdbaseTimerList,
-  MdbaseTimerReconciliation,
-  MutationEstimate,
-  MutationProgressState,
-  OperationRequestOptions,
-  PendingMutationSummary,
-  QueryInput,
-  QueryPage,
-  QueryPagesOptions,
-  QueryResult,
-  ReadInput,
-  ReadTypeInput,
-  RenameInput,
-  RenamePreflightResult,
-  RenameProgressOptions,
-  RenameResult,
-  UpdateInput,
-  UpdateTypeInput,
-  WatchOptions
-} from "./operation-types.js";
-import type {
-  MdbaseNativeNotificationRegistration,
-  MdbaseNativeNotificationRegistrationOptions,
-  MdbaseNotificationRegistration,
-  MdbaseNotificationRegistrationOptions
-} from "./notifications.js";
-import type {
-  MdbaseSelectionHistory,
-  MdbaseSessionSelection
-} from "./selection.js";
-import { DEFAULT_OPERATIONS } from "./internal-types.js";
-import type {
-  Application,
-  OperationAttempt,
-  PendingMutation,
-  StoredAuthorization,
-  StoredConnectionIndex,
-  StoredToken
-} from "./internal-types.js";
-import {
-  assertDeletePreview,
-  assertRenamePreview,
-  canonicalJson,
-  deleteEstimate,
-  directFallbackStatus,
-  isCancellation,
-  isMutation,
-  localNetworkPermission,
-  loopbackRequest,
-  operationFingerprint,
-  operationTransportError,
-  renameEstimate,
-  sameAuthorization,
-  throwIfCancelled,
-  uncertainDirectMutation,
-  uniqueOperations
-} from "./operation-helpers.js";
-import {
-  MemoryStorage,
-  apiError,
-  canonicalLoopbackUrl,
-  createPkce,
-  defaultCallbackUrl,
-  defaultManifestSource,
-  defaultRedirectUri,
-  defaultStorage,
-  isOpaquePortableManifest,
-  manifestStorageFingerprint,
-  oauthErrorCode,
-  parseDeviceAuthorization,
-  parseGrantScope,
-  parseStored,
-  stripTrailingSlash,
-  validAuthorityTokenResponse,
-  validStoredAuthority,
-  validStoredEncryption
-} from "./runtime-utils.js";
-import {
-  base64UrlBytes,
-  randomBase64Url
-} from "./base64.js";
-import type { MdbaseDeviceAuthorization } from "./authorization-types.js";
-
 export {
   decryptRelayResponse,
   encryptRelayRequest,
@@ -163,12 +8,17 @@ export {
   type GrantKeyRecord,
   type GrantKeyStore
 } from "./crypto.js";
-export * from "./collection-client.js";
+
 export * from "./authorization-types.js";
+export * from "./collection-client.js";
+export * from "./connect.js";
+export * from "./connection.js";
+export * from "./connection-types.js";
 export * from "./errors.js";
 export * from "./notifications.js";
 export * from "./operation-types.js";
 export * from "./selection.js";
+export * from "./session.js";
 export { createPkce } from "./runtime-utils.js";
 
 export type {
@@ -214,2570 +64,3 @@ export type {
   TypePackResourceDiff,
   TypePackSourceResource
 } from "@mdbase/connect-protocol";
-
-export interface MdbaseConnectOptions {
-  serverUrl: string;
-  /**
-   * A bundled v1 application manifest or its app-local URL. String values
-   * are loaded by this SDK and posted inline; Connect never fetches them.
-   */
-  manifest?: MdbaseAppManifest | string;
-  redirectUri?: string;
-  storage?: Storage;
-  /** Encrypted relay is required by default for newly authorized grants. */
-  relayEncryption?: "required" | "disabled";
-  keyStore?: GrantKeyStore;
-  /** Prefer same-computer connector access when the browser permits it. */
-  directAccess?: "auto" | "disabled";
-  /** Loopback origin override for development and automated testing. */
-  loopbackUrl?: string;
-  /** Override browser navigation, for example to use a native system browser. */
-  navigate?: (url: string) => void | Promise<void>;
-}
-
-export type MdbaseConnectionRoute = "remote" | "direct" | "relay";
-export type DirectAccessStatus =
-  | "disabled"
-  | "permission_required"
-  | "checking"
-  | "available"
-  | "unavailable"
-  | "denied";
-
-export interface MdbaseConnectionInfo {
-  collectionId: string;
-  displayName: string;
-  operations: CollectionOperation[];
-  scope: GrantScope;
-  route: MdbaseConnectionRoute;
-  directAccess: DirectAccessStatus;
-}
-
-export type MdbaseAuthorizationTarget =
-  | { kind: "choose" }
-  | { kind: "collection"; collectionId: string };
-
-export interface MdbaseAuthorizeOptions {
-  operations?: CollectionOperation[];
-  /** Choose any compatible collection, or require one exact collection. */
-  target?: MdbaseAuthorizationTarget;
-  /** App-local location to restore after the authorization callback. */
-  returnTo?: string;
-  /** Receives the short code even when the SDK also opens the approval page. */
-  onDeviceCode?: (authorization: MdbaseDeviceAuthorization) => void;
-  /** Replace the default popup for a downloaded application's approval page. */
-  openVerification?: (authorization: MdbaseDeviceAuthorization) => void | Promise<void>;
-  /** Stop polling and discard the unapproved, in-memory key. */
-  signal?: AbortSignal;
-}
-
-export interface MdbaseConnectionAuthorizeOptions {
-  operations?: CollectionOperation[];
-  returnTo?: string;
-  onDeviceCode?: (authorization: MdbaseDeviceAuthorization) => void;
-  openVerification?: (authorization: MdbaseDeviceAuthorization) => void | Promise<void>;
-  signal?: AbortSignal;
-}
-
-export interface MdbaseConnectEnvironment {
-  distribution: "web" | "portable";
-  applicationOrigin: string;
-  credentialStorage: "persistent" | "memory" | "custom";
-}
-
-export interface MdbaseAuthorizationResult<Frontmatter extends JsonObject = JsonObject> {
-  connection: MdbaseConnection<Frontmatter>;
-  returnTo?: string;
-}
-
-export type MdbaseAuthorizationOutcome<Frontmatter extends JsonObject = JsonObject> =
-  | { kind: "redirecting" }
-  | ({ kind: "connected" } & MdbaseAuthorizationResult<Frontmatter>);
-
-export interface MdbaseAuthorizationCapabilities {
-  authorized: boolean;
-  sufficient: boolean;
-  collectionId?: string;
-  grantedOperations: CollectionOperation[];
-  missingOperations: CollectionOperation[];
-}
-
-export interface MdbaseSyncTransport<Frontmatter extends JsonObject = JsonObject> {
-  openSession(): Promise<SyncSession>;
-  snapshot(snapshotId: string, page?: string): Promise<SyncSnapshotPage<Frontmatter>>;
-  changes(after: number, limit?: number): Promise<SyncChangesPage<Frontmatter>>;
-  mutate(mutation: SyncMutation): Promise<SyncMutationReceipt<Frontmatter>>;
-}
-
-export interface MdbaseSyncConnection<Frontmatter extends JsonObject = JsonObject> {
-  collectionId: string;
-  replicaId: string;
-  transport: MdbaseSyncTransport<Frontmatter>;
-}
-
-export class MdbaseConnect<Frontmatter extends JsonObject = JsonObject> {
-  private readonly internals: MdbaseConnectInternals<Frontmatter>;
-
-  constructor(options: MdbaseConnectOptions) {
-    this.internals = new MdbaseConnectInternals(options);
-  }
-
-  register(): Promise<Application> {
-    return this.internals.register();
-  }
-
-  authorize(options: MdbaseAuthorizeOptions = {}): Promise<MdbaseAuthorizationOutcome<Frontmatter>> {
-    return this.internals.authorize(options);
-  }
-
-  createSession(options: MdbaseSessionOptions): MdbaseSession<Frontmatter> {
-    return new MdbaseSession(this, options);
-  }
-
-  environment(): MdbaseConnectEnvironment {
-    return this.internals.environment();
-  }
-
-  completeAuthorization(
-    callbackUrl = defaultCallbackUrl()
-  ): Promise<MdbaseAuthorizationResult<Frontmatter>> {
-    return this.internals.completeAuthorization(callbackUrl);
-  }
-
-  connections(): MdbaseConnectionInfo[] {
-    return this.internals.connections();
-  }
-
-  connection(collectionId: string): MdbaseConnection<Frontmatter> | null {
-    return this.internals.connection(collectionId);
-  }
-
-  unavailableReason(collectionId: string): MdbaseUnavailableReason | null {
-    return this.internals.unavailableReason(collectionId);
-  }
-
-  onConnectionsChange(
-    listener: (connections: MdbaseConnectionInfo[]) => void
-  ): () => void {
-    return this.internals.onConnectionsChange(listener);
-  }
-
-  forgetAll(): void {
-    for (const connection of this.connections()) {
-      this.connection(connection.collectionId)?.forget();
-    }
-  }
-}
-
-export interface MdbaseSessionOptions {
-  selection: MdbaseSessionSelection;
-  operations?: CollectionOperation[];
-  autoSelect?: "only" | "never";
-}
-
-export type MdbaseUnavailableReason =
-  | "not_authorized"
-  | "authorization_lost"
-  | "invalid_stored_grant";
-
-export type MdbaseSessionSnapshot<Frontmatter extends JsonObject = JsonObject> =
-  | {
-      status: "unselected";
-      connections: MdbaseConnectionInfo[];
-    }
-  | {
-      status: "ready";
-      collectionId: string;
-      connection: MdbaseConnection<Frontmatter>;
-      info: MdbaseConnectionInfo;
-      access: MdbaseAuthorizationCapabilities;
-      connections: MdbaseConnectionInfo[];
-    }
-  | {
-      status: "unavailable";
-      collectionId: string;
-      reason: MdbaseUnavailableReason;
-      connections: MdbaseConnectionInfo[];
-    };
-
-export class MdbaseSession<Frontmatter extends JsonObject = JsonObject> {
-  private readonly operations: CollectionOperation[];
-  private readonly selection: MdbaseSessionSelection;
-  private readonly autoSelect: "only" | "never";
-  private readonly listeners = new Set<() => void>();
-  private snapshot: MdbaseSessionSnapshot<Frontmatter>;
-  private snapshotKey = "";
-  private stopConnections?: () => void;
-  private stopSelection?: () => void;
-  private stopActiveConnection?: () => void;
-  private activeCollectionId: string | null = null;
-  private unavailableReason: MdbaseUnavailableReason = "not_authorized";
-  private started = false;
-  private refreshing = false;
-  private refreshQueued = false;
-  private transactionDepth = 0;
-  private callbackPromise: Promise<MdbaseConnection<Frontmatter>> | null = null;
-
-  constructor(
-    private readonly connect: MdbaseConnect<Frontmatter>,
-    options: MdbaseSessionOptions
-  ) {
-    this.operations = uniqueOperations(options.operations ?? DEFAULT_OPERATIONS);
-    this.selection = options.selection;
-    this.autoSelect = options.autoSelect ?? "only";
-    this.snapshot = { status: "unselected", connections: [] };
-    this.refresh();
-  }
-
-  async start(): Promise<MdbaseSessionSnapshot<Frontmatter>> {
-    if (!this.started) {
-      this.started = true;
-      this.stopConnections = this.connect.onConnectionsChange(() => this.refresh());
-      this.stopSelection = this.selection.subscribe(() => {
-        this.unavailableReason = "not_authorized";
-        this.refresh();
-      });
-    }
-    const callback = this.selection.authorizationCallback();
-    if (callback) await this.handleAuthorizationCallback(callback);
-    else this.autoSelectOnlyConnection();
-    this.refresh();
-    return this.snapshot;
-  }
-
-  destroy(): void {
-    this.stopConnections?.();
-    this.stopSelection?.();
-    this.stopActiveConnection?.();
-    this.stopConnections = undefined;
-    this.stopSelection = undefined;
-    this.stopActiveConnection = undefined;
-    this.activeCollectionId = null;
-    this.started = false;
-  }
-
-  getSnapshot(): MdbaseSessionSnapshot<Frontmatter> {
-    return this.snapshot;
-  }
-
-  subscribe(listener: () => void): () => void {
-    this.listeners.add(listener);
-    return () => this.listeners.delete(listener);
-  }
-
-  select(
-    collectionId: string,
-    options: { history?: MdbaseSelectionHistory } = {}
-  ): MdbaseConnection<Frontmatter> {
-    const connection = this.connect.connection(collectionId);
-    if (!connection) {
-      throw new MdbaseConnectError(
-        "unknown_collection",
-        "This collection is not authorized for this application."
-      );
-    }
-    this.unavailableReason = "not_authorized";
-    this.selection.select(collectionId, options);
-    this.refresh();
-    return connection;
-  }
-
-  clearSelection(options: { history?: MdbaseSelectionHistory } = {}): void {
-    this.unavailableReason = "not_authorized";
-    this.selection.select(null, options);
-    this.refresh();
-  }
-
-  forget(collectionId: string): void {
-    this.transactionDepth += 1;
-    try {
-      const selected = this.selection.selectedCollectionId() === collectionId;
-      this.connect.connection(collectionId)?.forget();
-      if (selected) this.selection.select(null, { history: "replace" });
-    } finally {
-      this.transactionDepth -= 1;
-      this.refresh();
-    }
-  }
-
-  async authorize(
-    target: "choose" | "selected" | { collectionId: string },
-    options: Omit<MdbaseAuthorizeOptions, "operations" | "returnTo" | "target"> = {}
-  ): Promise<MdbaseAuthorizationOutcome<Frontmatter>> {
-    const selectedCollectionId = this.selection.selectedCollectionId();
-    if (target === "selected" && !selectedCollectionId) {
-      throw new MdbaseConnectError("collection_not_selected", "Choose a collection before updating its access.");
-    }
-    const targetCollectionId = typeof target === "object"
-      ? target.collectionId
-      : target === "selected"
-        ? selectedCollectionId!
-        : null;
-    this.transactionDepth += 1;
-    try {
-      const outcome = await this.connect.authorize({
-        ...options,
-        operations: this.operations,
-        target: targetCollectionId === null
-          ? { kind: "choose" }
-          : { kind: "collection", collectionId: targetCollectionId },
-        returnTo: this.selection.authorizationReturnTo()
-      });
-      if (outcome.kind === "connected") {
-        this.selection.finishAuthorization(outcome.returnTo, outcome.connection.collectionId);
-      }
-      return outcome;
-    } finally {
-      this.transactionDepth -= 1;
-      this.refresh();
-    }
-  }
-
-  async ensureOperations(
-    requiredOperations: CollectionOperation[]
-  ): Promise<MdbaseAuthorizationOutcome<Frontmatter> | { kind: "unchanged"; connection: MdbaseConnection<Frontmatter> }> {
-    const current = this.snapshot;
-    if (current.status !== "ready") {
-      throw new MdbaseConnectError("collection_not_ready", "Choose an authorized collection first.");
-    }
-    const required = uniqueOperations(requiredOperations);
-    const capabilities = current.connection.authorizationCapabilities(required);
-    if (capabilities.sufficient) return { kind: "unchanged", connection: current.connection };
-    this.transactionDepth += 1;
-    try {
-      const outcome = await this.connect.authorize({
-        operations: uniqueOperations([...capabilities.grantedOperations, ...capabilities.missingOperations]),
-        target: { kind: "collection", collectionId: current.collectionId },
-        returnTo: this.selection.authorizationReturnTo()
-      });
-      if (outcome.kind === "connected") {
-        this.selection.finishAuthorization(outcome.returnTo, outcome.connection.collectionId);
-      }
-      return outcome;
-    } finally {
-      this.transactionDepth -= 1;
-      this.refresh();
-    }
-  }
-
-  handleAuthorizationCallback(
-    callbackUrl: string
-  ): Promise<MdbaseConnection<Frontmatter>> {
-    if (!isAuthorizationCallbackUrl(callbackUrl)) {
-      return Promise.reject(new MdbaseConnectError("invalid_callback", "This URL is not an authorization callback."));
-    }
-    if (this.callbackPromise) return this.callbackPromise;
-    this.transactionDepth += 1;
-    const completion = this.connect.completeAuthorization(callbackUrl)
-      .then((result) => {
-        this.unavailableReason = "not_authorized";
-        this.selection.finishAuthorization(result.returnTo, result.connection.collectionId);
-        this.refresh();
-        return result.connection;
-      })
-      .catch((error) => {
-        const returnTo = authorizationReturnToFromError(error);
-        this.selection.clearAuthorizationCallback(returnTo);
-        this.refresh();
-        throw error;
-      })
-      .finally(() => {
-        this.transactionDepth -= 1;
-        this.refresh();
-        if (this.callbackPromise === completion) this.callbackPromise = null;
-      });
-    this.callbackPromise = completion;
-    return completion;
-  }
-
-  private autoSelectOnlyConnection(): void {
-    if (this.autoSelect !== "only" || this.selection.selectedCollectionId()) return;
-    const connections = this.connect.connections();
-    if (connections.length === 1) {
-      this.selection.select(connections[0].collectionId, { history: "replace" });
-    }
-  }
-
-  private refresh(): void {
-    if (this.transactionDepth > 0) {
-      this.refreshQueued = true;
-      return;
-    }
-    if (this.refreshing) {
-      this.refreshQueued = true;
-      return;
-    }
-    this.refreshing = true;
-    try {
-      do {
-        this.refreshQueued = false;
-        this.refreshOnce();
-      } while (this.refreshQueued);
-    } finally {
-      this.refreshing = false;
-    }
-  }
-
-  private refreshOnce(): void {
-    const collectionId = this.selection.selectedCollectionId();
-    const connections = this.connect.connections();
-    const connection = collectionId ? this.connect.connection(collectionId) : null;
-    if (connection?.collectionId !== this.activeCollectionId) {
-      this.stopActiveConnection?.();
-      this.activeCollectionId = connection?.collectionId ?? null;
-      this.stopActiveConnection = connection?.onConnectionChange((info) => {
-        if (!info && this.selection.selectedCollectionId() === connection.collectionId) {
-          this.unavailableReason = "authorization_lost";
-        }
-        this.refresh();
-      });
-    }
-
-    let next: MdbaseSessionSnapshot<Frontmatter>;
-    if (!collectionId) {
-      next = { status: "unselected", connections };
-    } else if (!connection || !connection.info()) {
-      next = {
-        status: "unavailable",
-        collectionId,
-        reason: this.connect.unavailableReason(collectionId) ?? this.unavailableReason,
-        connections
-      };
-    } else {
-      const info = connection.info()!;
-      next = {
-        status: "ready",
-        collectionId,
-        connection,
-        info,
-        access: connection.authorizationCapabilities(this.operations),
-        connections
-      };
-    }
-    const key = sessionSnapshotKey(next);
-    if (key === this.snapshotKey) return;
-    this.snapshot = next;
-    this.snapshotKey = key;
-    for (const listener of this.listeners) listener();
-  }
-}
-
-class MdbaseConnectInternals<Frontmatter extends JsonObject> {
-  readonly serverUrl: string;
-  readonly manifest: MdbaseAppManifest | string;
-  readonly manifestSource: string;
-  readonly redirectUri: string;
-  readonly storage: Storage;
-  readonly relayEncryption: "required" | "disabled";
-  readonly keyStore: GrantKeyStore;
-  readonly directAccessMode: "auto" | "disabled";
-  readonly loopbackUrl: string;
-  readonly navigate?: (url: string) => void | Promise<void>;
-  readonly credentialStorage: MdbaseConnectEnvironment["credentialStorage"];
-  private application: Application | null = null;
-  private readonly completionPromises = new Map<string, Promise<MdbaseAuthorizationResult<Frontmatter>>>();
-  private readonly connectionCache = new Map<string, MdbaseConnection<Frontmatter>>();
-  private readonly listeners = new Set<(connections: MdbaseConnectionInfo[]) => void>();
-  private readonly invalidatedConnections = new Map<string, MdbaseUnavailableReason>();
-
-  constructor(options: MdbaseConnectOptions) {
-    this.serverUrl = stripTrailingSlash(options.serverUrl);
-    this.manifest = options.manifest ?? defaultManifestSource();
-    this.manifestSource = typeof this.manifest === "string"
-      ? this.manifest
-      : this.manifest.distribution === "portable"
-        ? `bundle:${this.manifest.id}:${manifestStorageFingerprint(this.manifest)}`
-        : `bundle:${this.manifest.id}`;
-    const opaquePortable = isOpaquePortableManifest(this.manifest);
-    this.redirectUri = options.redirectUri ?? (
-      typeof this.manifest !== "string" && this.manifest.distribution === "portable"
-        ? ""
-        : defaultRedirectUri()
-    );
-    this.storage = options.storage ?? defaultStorage(opaquePortable);
-    this.relayEncryption = options.relayEncryption ?? "required";
-    this.keyStore = options.keyStore ?? (
-      opaquePortable ? new MemoryGrantKeyStore() : new IndexedDbGrantKeyStore()
-    );
-    this.credentialStorage = options.storage || options.keyStore
-      ? "custom"
-      : this.storage instanceof MemoryStorage ? "memory" : "persistent";
-    this.directAccessMode = options.directAccess ?? "auto";
-    this.loopbackUrl = canonicalLoopbackUrl(
-      options.loopbackUrl ?? `http://127.0.0.1:${DEFAULT_LOOPBACK_PORT}`
-    );
-    this.navigate = options.navigate;
-    if (typeof window !== "undefined" && this.storage === window.localStorage) {
-      window.addEventListener("storage", (event) => {
-        if (event.storageArea !== this.storage || !event.key?.startsWith(this.storagePrefix())) return;
-        this.connectionCache.get(collectionIdFromTokenKey(event.key))?.notifyStorageChanged();
-        this.emitConnections();
-      });
-    }
-  }
-
-  environment(): MdbaseConnectEnvironment {
-    const distribution = typeof this.manifest !== "string"
-      && this.manifest.distribution === "portable"
-      ? "portable"
-      : this.application?.distribution ?? "web";
-    return {
-      distribution,
-      applicationOrigin: distribution === "portable"
-        ? "null"
-        : this.defaultApplicationOrigin(),
-      credentialStorage: this.credentialStorage
-    };
-  }
-
-  async register(): Promise<Application> {
-    if (this.application) return this.application;
-    const response = await fetch(`${this.serverUrl}/v1/apps/register`, {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ manifest: await this.loadManifest() })
-    });
-    const body = await response.json();
-    if (!response.ok) throw apiError(body, "discovery_failed", "Application discovery failed.", response.status);
-    this.application = body.application;
-    return this.application!;
-  }
-
-  private async loadManifest(): Promise<MdbaseAppManifest> {
-    if (typeof this.manifest !== "string") return this.manifest;
-    const source = this.manifest;
-    let response: Response;
-    try {
-      response = await fetch(source, {
-        headers: { accept: "application/json" }
-      });
-    } catch (cause) {
-      throw new MdbaseConnectError(
-        "manifest_load_failed",
-        "The bundled application declaration could not be loaded.",
-        { cause }
-      );
-    }
-    if (!response.ok) {
-      throw new MdbaseConnectError(
-        "manifest_load_failed",
-        `The bundled application declaration returned HTTP ${response.status}.`,
-        { status: response.status }
-      );
-    }
-    try {
-      return await response.json() as MdbaseAppManifest;
-    } catch (cause) {
-      throw new MdbaseConnectError(
-        "invalid_application_manifest",
-        "The bundled application declaration is not valid JSON.",
-        { cause }
-      );
-    }
-  }
-
-  async authorize(
-    options: MdbaseAuthorizeOptions = {}
-  ): Promise<MdbaseAuthorizationOutcome<Frontmatter>> {
-    if (typeof location === "undefined" && !this.navigate && !options.openVerification) {
-      throw new MdbaseConnectError(
-        "browser_required",
-        "Authorization navigation requires a browser environment."
-      );
-    }
-    const portableDeclared = typeof this.manifest !== "string"
-      && this.manifest.distribution === "portable";
-    const popup = portableDeclared
-      && !options.openVerification
-      && typeof window !== "undefined"
-      ? window.open(
-          `${this.serverUrl}/device`,
-          "mdbase-connect-authorization",
-          "popup,width=620,height=760"
-        )
-      : null;
-    let application: Application;
-    try {
-      application = await this.register();
-    } catch (error) {
-      popup?.close();
-      throw error;
-    }
-    if (application.distribution === "portable") {
-      return this.authorizePortable(application, options, popup);
-    }
-    popup?.close();
-    const { verifier, challenge } = await createPkce();
-    const state = randomBase64Url(24);
-    const targetCollectionId = options.target?.kind === "collection"
-      ? options.target.collectionId
-      : undefined;
-    const keyHandle = this.relayEncryption === "required"
-      ? `grant:${application.id}:${state}`
-      : undefined;
-    const grantKey = keyHandle ? await this.keyStore.create(keyHandle) : undefined;
-    const pending: StoredAuthorization = {
-      version: 1,
-      verifier,
-      state,
-      clientId: application.id,
-      redirectUri: this.redirectUri,
-      relayEncryption: this.relayEncryption,
-      collectionId: targetCollectionId,
-      returnTo: options.returnTo,
-      keyHandle,
-      applicationAgreementPublicKey: grantKey?.agreementPublicKey,
-      applicationSigningPublicKey: grantKey?.signingPublicKey
-    };
-    this.storage.setItem(this.pendingKey(state), JSON.stringify(pending));
-    const authorize = new URL(`${this.serverUrl}/oauth/authorize`);
-    authorize.searchParams.set("client_id", application.id);
-    authorize.searchParams.set("redirect_uri", this.redirectUri);
-    authorize.searchParams.set("code_challenge", challenge);
-    authorize.searchParams.set("code_challenge_method", "S256");
-    authorize.searchParams.set("state", state);
-    authorize.searchParams.set(
-      "operations",
-      uniqueOperations(options.operations ?? DEFAULT_OPERATIONS).join(",")
-    );
-    if (targetCollectionId) {
-      authorize.searchParams.set("collection_id", targetCollectionId);
-    }
-    if (grantKey) {
-      authorize.searchParams.set("relay_protocol", "1");
-      authorize.searchParams.set(
-        "application_agreement_public_key",
-        grantKey.agreementPublicKey
-      );
-      authorize.searchParams.set(
-        "application_signing_public_key",
-        grantKey.signingPublicKey
-      );
-    }
-    if (this.navigate) await this.navigate(authorize.href);
-    else location.assign(authorize.href);
-    return { kind: "redirecting" };
-  }
-
-  private async authorizePortable(
-    application: Application,
-    options: MdbaseAuthorizeOptions,
-    popup: Window | null
-  ): Promise<MdbaseAuthorizationOutcome<Frontmatter>> {
-    if (this.relayEncryption !== "required") {
-      popup?.close();
-      throw new MdbaseConnectError(
-        "encryption_required",
-        "Downloaded applications require encrypted relay authorization."
-      );
-    }
-    const { verifier, challenge } = await createPkce();
-    const keyHandle = `grant:${application.id}:${randomBase64Url(24)}`;
-    const grantKey = await this.keyStore.create(keyHandle);
-    let response: Response;
-    let body: any;
-    try {
-      response = await fetch(`${this.serverUrl}/oauth/device_authorization`, {
-        method: "POST",
-        headers: { "content-type": "application/x-www-form-urlencoded" },
-        body: new URLSearchParams({
-          client_id: application.id,
-          operations: uniqueOperations(options.operations ?? DEFAULT_OPERATIONS).join(","),
-          ...(options.target?.kind === "collection"
-            ? { collection_id: options.target.collectionId }
-            : {}),
-          code_challenge: challenge,
-          code_challenge_method: "S256",
-          relay_protocol: "1",
-          application_agreement_public_key: grantKey.agreementPublicKey,
-          application_signing_public_key: grantKey.signingPublicKey
-        })
-      });
-      body = await response.json();
-    } catch (cause) {
-      popup?.close();
-      await this.keyStore.delete(keyHandle);
-      throw new MdbaseConnectError(
-        "device_authorization_failed",
-        "Downloaded application authorization could not be started.",
-        { cause }
-      );
-    }
-    if (!response.ok) {
-      popup?.close();
-      await this.keyStore.delete(keyHandle);
-      throw apiError(
-        body,
-        "device_authorization_failed",
-        "Downloaded application authorization could not be started.",
-        response.status
-      );
-    }
-    let authorization: MdbaseDeviceAuthorization;
-    try {
-      authorization = parseDeviceAuthorization(body);
-    } catch (error) {
-      popup?.close();
-      await this.keyStore.delete(keyHandle);
-      throw error;
-    }
-    options.onDeviceCode?.(authorization);
-    if (options.openVerification) {
-      await options.openVerification(authorization);
-    } else if (popup) {
-      try {
-        popup.location.href = authorization.verificationUriComplete;
-      } catch {
-        // The already-open verification page still accepts the displayed code.
-      }
-    } else {
-      await this.keyStore.delete(keyHandle);
-      throw new MdbaseConnectError(
-        "approval_window_blocked",
-        "The approval window was blocked. Show the provided verification link and code, then try again.",
-        { details: authorization }
-      );
-    }
-
-    let intervalSeconds = authorization.intervalSeconds;
-    try {
-      while (Date.now() < authorization.expiresAt) {
-        await abortableDelay(intervalSeconds * 1_000, options.signal);
-        if (options.signal?.aborted) {
-          throw new MdbaseConnectError(
-            "authorization_cancelled",
-            "Downloaded application authorization was cancelled."
-          );
-        }
-        let tokenResponse: Response;
-        try {
-          tokenResponse = await fetch(`${this.serverUrl}/oauth/token`, {
-            method: "POST",
-            headers: { "content-type": "application/x-www-form-urlencoded" },
-            body: new URLSearchParams({
-              grant_type: "urn:ietf:params:oauth:grant-type:device_code",
-              device_code: body.device_code,
-              client_id: application.id,
-              code_verifier: verifier
-            }),
-            signal: options.signal
-          });
-        } catch (cause) {
-          if (options.signal?.aborted) {
-            throw new MdbaseConnectError(
-              "authorization_cancelled",
-              "Downloaded application authorization was cancelled.",
-              { cause }
-            );
-          }
-          continue;
-        }
-        const tokenBody = await tokenResponse.json();
-        if (!tokenResponse.ok) {
-          const code = oauthErrorCode(tokenBody);
-          if (code === "authorization_pending") continue;
-          if (code === "slow_down") {
-            intervalSeconds += 5;
-            continue;
-          }
-          throw apiError(
-            tokenBody,
-            "token_exchange_failed",
-            "Downloaded application authorization could not be completed.",
-            tokenResponse.status
-          );
-        }
-        const authority = validAuthorityTokenResponse(tokenBody.authority, tokenBody.collection_id);
-        const localEncryption = tokenBody.encryption
-          && tokenBody.encryption.protocol_version === ENCRYPTED_RELAY_PROTOCOL_VERSION
-          && tokenBody.encryption.suite === RELAY_ENCRYPTION_SUITE
-          && tokenBody.encryption.application_agreement_public_key
-            === grantKey.agreementPublicKey;
-        if (tokenBody.application_origin !== "null") {
-          throw new MdbaseConnectError(
-            "invalid_token_response",
-            "Authorization did not bind the portable grant to its opaque application origin."
-          );
-        }
-        if (
-          tokenBody.authority
-          && (
-            !authority
-            || tokenBody.encryption != null
-            || tokenBody.authority.proof_public_key !== grantKey.signingPublicKey
-          )
-        ) {
-          throw new MdbaseConnectError(
-            "invalid_token_response",
-            "Authorization returned a remote authority capability that is not bound to this portable grant key."
-          );
-        }
-        if (!tokenBody.authority && !localEncryption) {
-          throw new MdbaseConnectError(
-            "encryption_required",
-            "Authorization did not establish the expected key-bound local portable grant."
-          );
-        }
-        const token = this.storeTokenResponse(
-          tokenBody,
-          application.id,
-          keyHandle
-        );
-        if (options.target?.kind === "collection"
-            && options.target.collectionId !== token.collectionId) {
-          this.removeToken(token.collectionId, token.keyHandle);
-          throw new MdbaseConnectError(
-            "collection_mismatch",
-            "The approved collection does not match the collection requested by this link."
-          );
-        }
-        popup?.close();
-        return {
-          kind: "connected",
-          connection: this.connection(token.collectionId)!,
-          ...(options.returnTo ? { returnTo: options.returnTo } : {})
-        };
-      }
-      throw new MdbaseConnectError(
-        "expired_token",
-        "The downloaded application authorization code expired."
-      );
-    } catch (error) {
-      popup?.close();
-      await this.keyStore.delete(keyHandle);
-      throw error;
-    }
-  }
-
-  completeAuthorization(callbackUrl: string): Promise<MdbaseAuthorizationResult<Frontmatter>> {
-    const state = new URL(callbackUrl).searchParams.get("state");
-    if (!state) {
-      return Promise.reject(new MdbaseConnectError(
-        "invalid_callback",
-        "Authorization callback is missing its state."
-      ));
-    }
-    const existing = this.completionPromises.get(state);
-    if (existing) return existing;
-    const completion = this.performAuthorizationCompletion(callbackUrl, state);
-    const shared = completion.finally(() => {
-      if (this.completionPromises.get(state) === shared) this.completionPromises.delete(state);
-    });
-    this.completionPromises.set(state, shared);
-    return shared;
-  }
-
-  private async performAuthorizationCompletion(
-    callbackUrl: string,
-    state: string
-  ): Promise<MdbaseAuthorizationResult<Frontmatter>> {
-    const callback = new URL(callbackUrl);
-    const code = callback.searchParams.get("code");
-    const pendingKey = this.pendingKey(state);
-    const pending = parseStored<StoredAuthorization>(this.storage.getItem(pendingKey));
-    if (!pending || pending.version !== 1 || state !== pending.state) {
-      throw new MdbaseConnectError("invalid_callback", "Authorization callback is missing or does not match this browser session.");
-    }
-    if (callback.searchParams.has("error")) {
-      if (pending.keyHandle) await this.keyStore.delete(pending.keyHandle);
-      this.storage.removeItem(pendingKey);
-      throw new MdbaseConnectError(
-        callback.searchParams.get("error") ?? "access_denied",
-        callback.searchParams.get("error_description") ?? "Collection access was not approved.",
-        { details: pending.returnTo ? { returnTo: pending.returnTo } : undefined }
-      );
-    }
-    if (!code) {
-      throw new MdbaseConnectError("invalid_callback", "Authorization callback is missing its code.");
-    }
-    const response = await fetch(`${this.serverUrl}/oauth/token`, {
-      method: "POST",
-      headers: { "content-type": "application/x-www-form-urlencoded" },
-      body: new URLSearchParams({
-        grant_type: "authorization_code",
-        code,
-        client_id: pending.clientId,
-        redirect_uri: pending.redirectUri,
-        code_verifier: pending.verifier
-      })
-    });
-    const body = await response.json();
-    if (!response.ok) throw apiError(body, "token_exchange_failed", "Authorization could not be completed.", response.status);
-    if (pending.relayEncryption === "required" && !body.authority && (
-      !body.encryption
-      || !pending.keyHandle
-      || body.encryption.application_agreement_public_key
-        !== pending.applicationAgreementPublicKey
-    )) {
-      if (pending.keyHandle) await this.keyStore.delete(pending.keyHandle);
-      this.storage.removeItem(pendingKey);
-      throw new MdbaseConnectError(
-        "encryption_required",
-        "Authorization did not establish the required encrypted relay grant."
-      );
-    }
-    if (body.authority?.proof_public_key) {
-      if (
-        !pending.keyHandle
-        || !pending.applicationSigningPublicKey
-        || body.authority.proof_public_key !== pending.applicationSigningPublicKey
-      ) {
-        if (pending.keyHandle) await this.keyStore.delete(pending.keyHandle);
-        this.storage.removeItem(pendingKey);
-        throw new MdbaseConnectError(
-          "invalid_token_response",
-          "Authorization returned a remote authority capability bound to another application key."
-        );
-      }
-    } else if (body.authority && pending.keyHandle) {
-      await this.keyStore.delete(pending.keyHandle);
-    }
-    const token = this.storeTokenResponse(
-      body,
-      pending.clientId,
-      body.authority?.proof_public_key ? pending.keyHandle : body.authority ? undefined : pending.keyHandle
-    );
-    if (pending.collectionId && pending.collectionId !== token.collectionId) {
-      this.removeToken(token.collectionId, token.keyHandle);
-      this.storage.removeItem(pendingKey);
-      throw new MdbaseConnectError(
-        "collection_mismatch",
-        "The approved collection does not match the collection requested by this link."
-      );
-    }
-    this.storage.removeItem(pendingKey);
-    return {
-      connection: this.connection(token.collectionId)!,
-      ...(pending.returnTo ? { returnTo: pending.returnTo } : {})
-    };
-  }
-
-  connections(): MdbaseConnectionInfo[] {
-    const connections: MdbaseConnectionInfo[] = [];
-    for (const collectionId of this.connectionIds()) {
-      const info = this.connection(collectionId)?.info();
-      if (info) connections.push(info);
-    }
-    return connections.sort((left, right) =>
-      left.displayName.localeCompare(right.displayName) || left.collectionId.localeCompare(right.collectionId)
-    );
-  }
-
-  connection(collectionId: string): MdbaseConnection<Frontmatter> | null {
-    if (!this.storage.getItem(this.tokenKey(collectionId))) return null;
-    let connection = this.connectionCache.get(collectionId);
-    if (!connection) {
-      connection = new MdbaseConnection(this, collectionId);
-      this.connectionCache.set(collectionId, connection);
-    }
-    return connection.info() ? connection : null;
-  }
-
-  unavailableReason(collectionId: string): MdbaseUnavailableReason | null {
-    return this.invalidatedConnections.get(collectionId) ?? null;
-  }
-
-  onConnectionsChange(listener: (connections: MdbaseConnectionInfo[]) => void): () => void {
-    this.listeners.add(listener);
-    listener(this.connections());
-    return () => this.listeners.delete(listener);
-  }
-
-  storeTokenResponse(body: any, clientId: string, keyHandle?: string): StoredToken {
-    const collectionId = body.collection_id;
-    if (typeof collectionId !== "string") {
-      throw new MdbaseConnectError("invalid_token_response", "Authorization returned no collection ID.");
-    }
-    const scope = parseGrantScope(body.scope);
-    if (!scope) {
-      throw new MdbaseConnectError(
-        "invalid_token_response",
-        "Authorization returned no valid collection scope."
-      );
-    }
-    if (body.authority && !validAuthorityTokenResponse(body.authority, collectionId)) {
-      throw new MdbaseConnectError(
-        "invalid_token_response",
-        "Authorization returned an invalid remote authority capability."
-      );
-    }
-    if (body.authority && body.encryption) {
-      throw new MdbaseConnectError(
-        "invalid_token_response",
-        "Authorization returned conflicting collection transports."
-      );
-    }
-    if (body.encryption) {
-      try {
-        validateGrantEncryption(body.encryption);
-      } catch {
-        throw new MdbaseConnectError(
-          "invalid_token_response",
-          "Authorization returned an invalid encrypted relay binding."
-        );
-      }
-      if (
-        body.encryption.collection_id !== collectionId
-        || typeof body.grant_id !== "string"
-        || body.grant_id.length === 0
-      ) {
-        throw new MdbaseConnectError(
-          "invalid_token_response",
-          "Authorization returned an encrypted relay binding for another grant."
-        );
-      }
-    }
-    const previous = parseStored<StoredToken>(this.storage.getItem(this.tokenKey(collectionId)));
-    if (previous?.keyHandle && previous.keyHandle !== keyHandle) {
-      void this.keyStore.delete(previous.keyHandle).catch(() => undefined);
-    }
-    const token: StoredToken = {
-      version: 1,
-      accessToken: body.access_token,
-      refreshToken: body.refresh_token,
-      clientId,
-      collectionId,
-      collectionName: body.collection_name ?? `Collection ${collectionId.slice(0, 8)}`,
-      operations: body.operations,
-      scope,
-      expiresAt: Date.now() + body.expires_in * 1_000,
-      refreshExpiresAt: body.refresh_expires_in
-        ? Date.now() + body.refresh_expires_in * 1_000
-        : undefined,
-      grantId: body.grant_id,
-      encryption: body.encryption ?? undefined,
-      applicationOrigin: body.application_origin ?? this.defaultApplicationOrigin(),
-      keyHandle,
-      savedAt: Date.now(),
-      authority: body.authority ? {
-        operationsUrl: body.authority.operations_url,
-        syncUrl: body.authority.sync_url,
-        replicaId: body.authority.replica_id,
-        accessToken: body.authority.access_token,
-        proofPublicKey: body.authority.proof_public_key
-      } : undefined
-    };
-    this.storage.setItem(this.tokenKey(collectionId), JSON.stringify(token));
-    this.addConnectionId(collectionId);
-    this.invalidatedConnections.delete(collectionId);
-    this.connectionCache.delete(collectionId);
-    this.emitConnections();
-    return token;
-  }
-
-  removeToken(
-    collectionId: string,
-    keyHandle?: string,
-    reason: MdbaseUnavailableReason = "authorization_lost"
-  ): void {
-    this.invalidatedConnections.set(collectionId, reason);
-    if (keyHandle) void this.keyStore.delete(keyHandle).catch(() => undefined);
-    this.storage.removeItem(this.tokenKey(collectionId));
-    this.storage.removeItem(this.pendingMutationKey(collectionId));
-    for (const transport of ["web_push", "fcm"] as const) {
-      this.storage.removeItem(this.notificationKey(collectionId, transport));
-    }
-    this.storage.setItem(
-      this.connectionsKey(),
-      JSON.stringify({
-        version: 1,
-        collectionIds: this.connectionIds().filter((id) => id !== collectionId)
-      } satisfies StoredConnectionIndex)
-    );
-    this.connectionCache.delete(collectionId);
-    this.emitConnections();
-  }
-
-  tokenKey(collectionId: string): string {
-    return `${this.storagePrefix()}:token:${collectionId}`;
-  }
-
-  pendingMutationKey(collectionId: string): string {
-    return `${this.storagePrefix()}:pending-mutation:${collectionId}`;
-  }
-
-  notificationKey(collectionId: string, transport: "web_push" | "fcm" = "web_push"): string {
-    return `${this.storagePrefix()}:notifications:${collectionId}:${transport}`;
-  }
-
-  directPreferenceKey(): string {
-    return `mdbase-connect:direct:${this.defaultApplicationOrigin()}`;
-  }
-
-  defaultApplicationOrigin(): string {
-    if (
-      (typeof this.manifest !== "string" && this.manifest.distribution === "portable")
-      || this.application?.distribution === "portable"
-    ) {
-      return "null";
-    }
-    const redirect = new URL(this.redirectUri);
-    if (["http:", "https:"].includes(redirect.protocol)) return redirect.origin;
-    if (typeof location !== "undefined") return location.origin;
-    if (
-      this.manifest
-      && typeof this.manifest !== "string"
-    ) {
-      return new URL(this.manifest.homepage).origin;
-    }
-    try {
-      return new URL(this.manifestSource).origin;
-    } catch {
-      return "";
-    }
-  }
-
-  private pendingKey(state: string): string {
-    return `${this.storagePrefix()}:pending:${state}`;
-  }
-
-  private storagePrefix(): string {
-    return `mdbase-connect:${this.serverUrl}:${this.manifestSource}`;
-  }
-
-  private connectionsKey(): string {
-    return `${this.storagePrefix()}:connections`;
-  }
-
-  private connectionIds(): string[] {
-    const index = parseStored<StoredConnectionIndex>(this.storage.getItem(this.connectionsKey()));
-    return index?.version === 1 && Array.isArray(index.collectionIds)
-      && index.collectionIds.every((collectionId) => typeof collectionId === "string")
-      ? index.collectionIds
-      : [];
-  }
-
-  private addConnectionId(collectionId: string): void {
-    this.storage.setItem(
-      this.connectionsKey(),
-      JSON.stringify({
-        version: 1,
-        collectionIds: [...new Set([...this.connectionIds(), collectionId])]
-      } satisfies StoredConnectionIndex)
-    );
-  }
-
-  private emitConnections(): void {
-    const connections = this.connections();
-    for (const listener of this.listeners) listener(connections);
-  }
-}
-
-function collectionIdFromTokenKey(key: string): string {
-  const marker = ":token:";
-  const index = key.lastIndexOf(marker);
-  return index < 0 ? "" : key.slice(index + marker.length);
-}
-
-export class MdbaseConnection<Frontmatter extends JsonObject = JsonObject> {
-  private readonly serverUrl: string;
-  private readonly storage: Storage;
-  private readonly keyStore: GrantKeyStore;
-  private readonly directAccessMode: "auto" | "disabled";
-  private readonly loopbackUrl: string;
-  private refreshPromise: Promise<StoredToken> | null = null;
-  private readonly collectionClient: MdbaseCollectionClient<Frontmatter>;
-  private directStatus: DirectAccessStatus;
-  private currentRoute: MdbaseConnectionRoute = "relay";
-  private directFailures = 0;
-  private directRetryAt = 0;
-  private readonly connectionListeners = new Set<(connection: MdbaseConnectionInfo | null) => void>();
-
-  constructor(
-    private readonly internals: MdbaseConnectInternals<Frontmatter>,
-    readonly collectionId: string
-  ) {
-    this.serverUrl = internals.serverUrl;
-    this.storage = internals.storage;
-    this.keyStore = internals.keyStore;
-    this.directAccessMode = internals.directAccessMode;
-    this.loopbackUrl = internals.loopbackUrl;
-    this.directStatus = this.directAccessMode === "disabled" ? "disabled" : "unavailable";
-    this.collectionClient = new MdbaseCollectionClient({
-      operation: (operation, input, requestOptions) => this.performOperation(operation, input, requestOptions)
-    });
-  }
-
-  get displayName(): string {
-    return this.currentToken()?.collectionName ?? `Collection ${this.collectionId.slice(0, 8)}`;
-  }
-
-  get operations(): CollectionOperation[] {
-    return [...(this.currentToken()?.operations ?? [])];
-  }
-
-  get scope(): GrantScope {
-    const scope = this.currentToken()?.scope;
-    if (!scope) {
-      throw new MdbaseConnectError(
-        "not_authorized",
-        "This collection is no longer authorized for this application."
-      );
-    }
-    return scope;
-  }
-
-  get directAccess(): DirectAccessStatus {
-    return this.currentToken()?.authority ? "disabled" : this.directStatus;
-  }
-
-  get route(): MdbaseConnectionRoute {
-    return this.currentToken()?.authority ? "remote" : this.currentRoute;
-  }
-
-  register(): Promise<Application> {
-    return this.internals.register();
-  }
-
-  info(): MdbaseConnectionInfo | null {
-    const token = this.currentToken();
-    return token ? {
-      collectionId: token.collectionId,
-      displayName: token.collectionName,
-      operations: [...token.operations],
-      scope: token.scope,
-      route: token.authority ? "remote" : this.currentRoute,
-      directAccess: token.authority ? "disabled" : this.directStatus
-    } : null;
-  }
-
-  authorizationCapabilities(
-    requiredOperations: CollectionOperation[] = DEFAULT_OPERATIONS
-  ): MdbaseAuthorizationCapabilities {
-    const grantedOperations = this.operations;
-    const missingOperations = uniqueOperations(requiredOperations)
-      .filter((operation) => !grantedOperations.includes(operation));
-    return {
-      authorized: this.info() !== null,
-      sufficient: this.info() !== null && missingOperations.length === 0,
-      collectionId: this.collectionId,
-      grantedOperations,
-      missingOperations
-    };
-  }
-
-  hasOperations(requiredOperations: CollectionOperation[]): boolean {
-    return this.authorizationCapabilities(requiredOperations).sufficient;
-  }
-
-  authorize(
-    options: MdbaseConnectionAuthorizeOptions = {}
-  ): Promise<MdbaseAuthorizationOutcome<Frontmatter>> {
-    return this.internals.authorize({
-      ...options,
-      target: { kind: "collection", collectionId: this.collectionId }
-    });
-  }
-
-  async requestOperations(
-    requiredOperations: CollectionOperation[],
-    options: Pick<MdbaseConnectionAuthorizeOptions, "returnTo"> = {}
-  ): Promise<
-    MdbaseAuthorizationOutcome<Frontmatter>
-    | { kind: "unchanged"; connection: MdbaseConnection<Frontmatter> }
-  > {
-    const capabilities = this.authorizationCapabilities(requiredOperations);
-    if (capabilities.sufficient) return { kind: "unchanged", connection: this };
-    return this.authorize({
-      ...options,
-      operations: uniqueOperations([
-        ...capabilities.grantedOperations,
-        ...capabilities.missingOperations
-      ])
-    });
-  }
-
-  onConnectionChange(listener: (connection: MdbaseConnectionInfo | null) => void): () => void {
-    this.connectionListeners.add(listener);
-    listener(this.info());
-    return () => this.connectionListeners.delete(listener);
-  }
-
-  notifyStorageChanged(): void {
-    this.emitConnection();
-  }
-
-  async checkDirectAccess(): Promise<DirectAccessStatus> {
-    const token = this.currentToken();
-    if (!this.directEligible(token)) return this.setDirectStatus("disabled");
-    const permission = await localNetworkPermission();
-    if (permission === "denied") return this.setDirectStatus("denied");
-    if ((permission === "prompt" || permission === null)
-        && this.storage.getItem(this.directPreferenceKey()) !== "enabled") {
-      return this.setDirectStatus("permission_required");
-    }
-    return this.probeDirectAccess();
-  }
-
-  /** Call from a user gesture to request browser permission for direct local access. */
-  async requestDirectAccess(): Promise<DirectAccessStatus> {
-    const token = this.currentToken();
-    if (!this.directCapable(token)) return this.setDirectStatus("disabled");
-    this.storage.setItem(this.directPreferenceKey(), "enabled");
-    this.directRetryAt = 0;
-    return this.probeDirectAccess();
-  }
-
-  disableDirectAccess(): void {
-    this.storage.setItem(this.directPreferenceKey(), "disabled");
-    this.setDirectStatus("disabled");
-    this.setRoute("relay");
-  }
-
-  /**
-   * Return one offline-replication transport regardless of whether the
-   * authority is remote, directly connected, or relay connected.
-   * Authority credentials and routing stay private inside the SDK.
-   */
-  sync(): MdbaseSyncConnection<Frontmatter> | null {
-    const token = this.currentToken();
-    if (!token) return null;
-    const collectionId = token.collectionId;
-    if (!token.authority) {
-      if (!token.grantId || !token.operations.includes("sync")) return null;
-      const replicaId = token.grantId;
-      return {
-        collectionId,
-        replicaId,
-        transport: {
-          openSession: () => this.performOperation("sync", { action: "open_session" }),
-          snapshot: (snapshotId, page) => this.performOperation("sync", {
-            action: "snapshot",
-            snapshot_id: snapshotId,
-            ...(page ? { page } : {})
-          }),
-          changes: (after, limit = 200) => this.performOperation("sync", {
-            action: "changes",
-            after,
-            limit
-          }),
-          mutate: (mutation) => this.performOperation("sync", {
-            action: "mutate",
-            mutation
-          })
-        }
-      };
-    }
-    const replicaId = token.authority.replicaId;
-    return {
-      collectionId,
-      replicaId,
-      transport: {
-        openSession: () => this.performAuthoritySyncRequest(collectionId, replicaId, "POST", "sessions"),
-        snapshot: (snapshotId, page) => {
-          const query = new URLSearchParams({ snapshot_id: snapshotId });
-          if (page) query.set("page", page);
-          return this.performAuthoritySyncRequest(collectionId, replicaId, "GET", `snapshot?${query}`);
-        },
-        changes: (after, limit = 200) => this.performAuthoritySyncRequest(
-          collectionId,
-          replicaId,
-          "GET",
-          `changes?${new URLSearchParams({ after: String(after), limit: String(limit) })}`
-        ),
-        mutate: (mutation) => this.performAuthoritySyncRequest(
-          collectionId,
-          replicaId,
-          "POST",
-          "mutations",
-          mutation
-        )
-      }
-    };
-  }
-
-  /**
-   * Register this browser installation for manifest-declared Web Push.
-   *
-   * Criteria are evaluated by the collection authority. Push payloads contain
-   * only an opaque cursor and static presentation copy.
-   */
-  async registerNotifications(
-    options: MdbaseNotificationRegistrationOptions
-  ): Promise<MdbaseNotificationRegistration> {
-    const token = await this.authorizedToken();
-    if (!token) {
-      throw new MdbaseConnectError(
-        "not_authorized",
-        "Connect this application before enabling notifications."
-      );
-    }
-    const application = await this.register();
-    const declared = application.notifications?.criteria.map((criterion) => criterion.id) ?? [];
-    const criteria = [...new Set(options.criteria ?? declared)];
-    const undeclared = criteria.find((criterion) => !declared.includes(criterion));
-    if (undeclared) {
-      throw new MdbaseConnectError(
-        "notification_criterion_not_declared",
-        `The application manifest does not declare notification criterion ${undeclared}.`
-      );
-    }
-    if (criteria.length === 0) {
-      throw new MdbaseConnectError(
-        "notifications_not_declared",
-        "This application manifest does not declare any notification criteria."
-      );
-    }
-    const keyResponse = await fetch(`${this.serverUrl}/v1/notifications/vapid-public-key`);
-    const keyBody = await keyResponse.json();
-    if (!keyResponse.ok) {
-      throw apiError(keyBody, "notifications_unavailable", "Push notifications are unavailable.", keyResponse.status);
-    }
-    let pushSubscription = await options.serviceWorker.pushManager.getSubscription();
-    if (!pushSubscription) {
-      pushSubscription = await options.serviceWorker.pushManager.subscribe({
-        userVisibleOnly: true,
-        applicationServerKey: base64UrlBytes(keyBody.public_key)
-      });
-    }
-    const serialized = pushSubscription.toJSON();
-    if (!serialized.endpoint || !serialized.keys?.p256dh || !serialized.keys.auth) {
-      throw new MdbaseConnectError(
-        "invalid_push_subscription",
-        "The browser returned an incomplete push subscription."
-      );
-    }
-    const previous = parseStored<MdbaseNotificationRegistration>(
-      this.storage.getItem(this.notificationKey())
-    );
-    const installationId = options.installationId
-      ?? previous?.installationId
-      ?? randomBase64Url(24);
-    const channelResponse = await fetch(`${this.serverUrl}/v1/notifications/channels`, {
-      method: "POST",
-      headers: {
-        authorization: `Bearer ${token.accessToken}`,
-        "content-type": "application/json"
-      },
-      body: JSON.stringify({
-        installation_id: installationId,
-        criteria,
-        subscription: {
-          endpoint: serialized.endpoint,
-          expirationTime: serialized.expirationTime ?? null,
-          keys: serialized.keys
-        }
-      })
-    });
-    const channelBody = await channelResponse.json();
-    if (!channelResponse.ok) {
-      throw apiError(channelBody, "notification_registration_failed", "Could not register push notifications.", channelResponse.status);
-    }
-    if (previous?.channelId && previous.channelId !== channelBody.channel_id) {
-      void fetch(`${this.serverUrl}/v1/notifications/channels/${previous.channelId}`, {
-        method: "DELETE",
-        headers: { authorization: `Bearer ${token.accessToken}` }
-      }).catch(() => undefined);
-    }
-    const registration = {
-      channelId: channelBody.channel_id,
-      installationId,
-      criteria
-    };
-    this.storage.setItem(this.notificationKey(), JSON.stringify(registration));
-    return registration;
-  }
-
-  /**
-   * Register an iOS or Android installation for Connect-managed FCM.
-   *
-   * The application manifest selects the Firebase project. Re-register the
-   * same installation whenever Firebase refreshes its token.
-   */
-  async registerNativeNotifications(
-    options: MdbaseNativeNotificationRegistrationOptions
-  ): Promise<MdbaseNativeNotificationRegistration> {
-    const token = await this.authorizedToken();
-    if (!token) {
-      throw new MdbaseConnectError(
-        "not_authorized",
-        "Connect this application before enabling notifications."
-      );
-    }
-    const application = await this.register();
-    if (application.notifications?.native_delivery?.mode !== "managed_fcm") {
-      throw new MdbaseConnectError(
-        "managed_fcm_not_declared",
-        "This application does not declare Connect-managed native notifications."
-      );
-    }
-    const declared = application.notifications.criteria.map(
-      (criterion) => criterion.id
-    );
-    const criteria = [...new Set(options.criteria ?? declared)];
-    const undeclared = criteria.find((criterion) => !declared.includes(criterion));
-    if (undeclared) {
-      throw new MdbaseConnectError(
-        "notification_criterion_not_declared",
-        `The application manifest does not declare notification criterion ${undeclared}.`
-      );
-    }
-    if (criteria.length === 0) {
-      throw new MdbaseConnectError(
-        "notifications_not_declared",
-        "This application manifest does not declare any notification criteria."
-      );
-    }
-    const storageKey = this.notificationKey("fcm");
-    const previous = parseStored<MdbaseNativeNotificationRegistration>(
-      this.storage.getItem(storageKey)
-    );
-    const installationId = options.installationId
-      ?? previous?.installationId
-      ?? randomBase64Url(24);
-    const response = await fetch(`${this.serverUrl}/v1/notifications/channels`, {
-      method: "POST",
-      headers: {
-        authorization: `Bearer ${token.accessToken}`,
-        "content-type": "application/json"
-      },
-      body: JSON.stringify({
-        installation_id: installationId,
-        criteria,
-        transport: "fcm",
-        token: options.token
-      })
-    });
-    const body = await response.json();
-    if (!response.ok) {
-      throw apiError(
-        body,
-        "notification_registration_failed",
-        "Could not register native notifications.",
-        response.status
-      );
-    }
-    if (previous?.channelId && previous.channelId !== body.channel_id) {
-      void this.deleteNotificationChannel(previous.channelId, token.accessToken)
-        .catch(() => undefined);
-    }
-    const registration: MdbaseNativeNotificationRegistration = {
-      channelId: body.channel_id,
-      installationId,
-      transport: "fcm",
-      criteria
-    };
-    this.storage.setItem(storageKey, JSON.stringify(registration));
-    return registration;
-  }
-
-  async unregisterNativeNotifications(): Promise<void> {
-    const storageKey = this.notificationKey("fcm");
-    const registration = parseStored<MdbaseNativeNotificationRegistration>(
-      this.storage.getItem(storageKey)
-    );
-    const token = await this.authorizedToken();
-    if (registration?.channelId && !token) {
-      throw new MdbaseConnectError(
-        "not_authorized",
-        "Reconnect this application before disabling native notifications."
-      );
-    }
-    if (registration?.channelId && token) {
-      await this.deleteNotificationChannel(
-        registration.channelId,
-        token.accessToken
-      );
-    }
-    this.storage.removeItem(storageKey);
-  }
-
-  async unregisterNotifications(
-    serviceWorker?: ServiceWorkerRegistration
-  ): Promise<void> {
-    const registration = parseStored<MdbaseNotificationRegistration>(
-      this.storage.getItem(this.notificationKey())
-    );
-    const token = await this.authorizedToken();
-    if (registration?.channelId && !token) {
-      throw new MdbaseConnectError(
-        "not_authorized",
-        "Reconnect this application before disabling push notifications."
-      );
-    }
-    if (registration?.channelId && token) {
-      const response = await fetch(`${this.serverUrl}/v1/notifications/channels/${registration.channelId}`, {
-        method: "DELETE",
-        headers: { authorization: `Bearer ${token.accessToken}` }
-      });
-      if (!response.ok && response.status !== 404) {
-        const body = await response.json();
-        throw apiError(
-          body,
-          "notification_unregistration_failed",
-          "Could not unregister push notifications.",
-          response.status
-        );
-      }
-    }
-    this.storage.removeItem(this.notificationKey());
-    const subscription = await serviceWorker?.pushManager.getSubscription();
-    await subscription?.unsubscribe();
-  }
-
-  forget(): void {
-    const token = this.currentToken();
-    this.internals.removeToken(this.collectionId, token?.keyHandle, "not_authorized");
-    this.setRoute("relay");
-    this.emitConnection();
-  }
-
-  describe(): Promise<CollectionDescription> {
-    return this.collectionClient.describe();
-  }
-
-  changes(input: ChangesInput = {}, options?: OperationRequestOptions): Promise<CollectionChangesPage> {
-    return this.collectionClient.changes(input, options);
-  }
-
-  read(input: ReadInput): Promise<MdbaseOperationEnvelope<RecordDocument<Frontmatter>>> {
-    return this.collectionClient.read(input);
-  }
-
-  query(input: QueryInput = {}, options?: OperationRequestOptions): Promise<MdbaseOperationEnvelope<QueryResult<Frontmatter>>> {
-    return this.collectionClient.query(input, options);
-  }
-
-  queryPages(input: QueryInput = {}, options: QueryPagesOptions<Frontmatter> = {}): AsyncGenerator<QueryPage<Frontmatter>> {
-    return this.collectionClient.queryPages(input, options);
-  }
-
-  queryAll(input: QueryInput = {}, options: QueryPagesOptions<Frontmatter> = {}): Promise<QueryResult<Frontmatter>> {
-    return this.collectionClient.queryAll(input, options);
-  }
-
-  listViews(): Promise<MdbaseOperationEnvelope<SavedViewList>> {
-    return this.collectionClient.listViews();
-  }
-
-  executeView(input: ExecuteViewInput): Promise<MdbaseOperationEnvelope<SavedViewExecution<Frontmatter>>> {
-    return this.collectionClient.executeView(input);
-  }
-
-  readViewSource(input: ReadViewSourceInput): Promise<MdbaseOperationEnvelope<SavedViewSourceDocument>> {
-    return this.collectionClient.readViewSource(input);
-  }
-
-  createViewSource(input: CreateViewSourceInput): Promise<MdbaseOperationEnvelope<SavedViewSourceDocument>> {
-    return this.collectionClient.createViewSource(input);
-  }
-
-  updateViewSource(input: UpdateViewSourceInput): Promise<MdbaseOperationEnvelope<SavedViewSourceDocument>> {
-    return this.collectionClient.updateViewSource(input);
-  }
-
-  deleteViewSource(input: DeleteViewSourceInput): Promise<MdbaseOperationEnvelope<DeleteViewSourceResult>> {
-    return this.collectionClient.deleteViewSource(input);
-  }
-
-  create(input: CreateInput<Frontmatter>): Promise<MdbaseOperationEnvelope<RecordDocument<Frontmatter>>> {
-    return this.collectionClient.create(input);
-  }
-
-  update(input: UpdateInput<Frontmatter>): Promise<MdbaseOperationEnvelope<RecordDocument<Frontmatter>>> {
-    return this.collectionClient.update(input);
-  }
-
-  delete(input: DeleteInput, options?: OperationRequestOptions): Promise<MdbaseOperationEnvelope<DeleteResult>> {
-    return this.collectionClient.delete(input, options);
-  }
-
-  preflightDelete(input: DeleteInput, options?: OperationRequestOptions): Promise<MdbaseOperationEnvelope<DeletePreflightResult>> {
-    return this.collectionClient.preflightDelete(input, options);
-  }
-
-  rename(input: RenameInput, options?: OperationRequestOptions): Promise<MdbaseOperationEnvelope<RenameResult>> {
-    return this.collectionClient.rename(input, options);
-  }
-
-  preflightRename(input: RenameInput, options?: OperationRequestOptions): Promise<MdbaseOperationEnvelope<RenamePreflightResult>> {
-    return this.collectionClient.preflightRename(input, options);
-  }
-
-  async renameWithProgress(
-    input: RenameInput,
-    options: RenameProgressOptions = {}
-  ): Promise<MdbaseOperationEnvelope<RenameResult>> {
-    const started = Date.now();
-    const resumed = this.pendingMutation()?.operation === "rename";
-    let estimate: MutationEstimate | undefined;
-    const emit = (state: MutationProgressState, cancellable: boolean, completedUnits = 0) => {
-      options.onProgress?.({
-        operation: "rename",
-        state,
-        elapsedMs: Date.now() - started,
-        cancellable,
-        resumed,
-        completedUnits,
-        ...(estimate ? { estimate } : {})
-      });
-    };
-    try {
-      throwIfCancelled(options.signal);
-      emit("preflighting", true);
-      const preview = options.preflight ?? unwrapOperation(await this.preflightRename(input, {
-        signal: options.signal
-      }));
-      assertRenamePreview(input, preview);
-      estimate = renameEstimate(input, preview);
-      emit("ready", true);
-      throwIfCancelled(options.signal);
-      const cancellable = this.hasResumableMutationTransport();
-      emit("applying", cancellable);
-      const result = await this.rename(input, cancellable ? { signal: options.signal } : undefined);
-      emit("completed", false, estimate.totalUnits);
-      return result;
-    } catch (error) {
-      if (isCancellation(error, options.signal)) emit("cancelled", false);
-      throw error;
-    }
-  }
-
-  async deleteWithProgress(
-    input: DeleteInput,
-    options: DeleteProgressOptions = {}
-  ): Promise<MdbaseOperationEnvelope<DeleteResult>> {
-    const started = Date.now();
-    const resumed = this.pendingMutation()?.operation === "delete";
-    let estimate: MutationEstimate | undefined;
-    const emit = (state: MutationProgressState, cancellable: boolean, completedUnits = 0) => {
-      options.onProgress?.({
-        operation: "delete",
-        state,
-        elapsedMs: Date.now() - started,
-        cancellable,
-        resumed,
-        completedUnits,
-        ...(estimate ? { estimate } : {})
-      });
-    };
-    try {
-      throwIfCancelled(options.signal);
-      emit("preflighting", true);
-      const preview = options.preflight ?? unwrapOperation(await this.preflightDelete(input, {
-        signal: options.signal
-      }));
-      assertDeletePreview(input, preview);
-      estimate = deleteEstimate(preview);
-      emit("ready", true);
-      throwIfCancelled(options.signal);
-      const cancellable = this.hasResumableMutationTransport();
-      emit("applying", cancellable);
-      const result = await this.delete(input, cancellable ? { signal: options.signal } : undefined);
-      emit("completed", false, estimate.totalUnits);
-      return result;
-    } catch (error) {
-      if (isCancellation(error, options.signal)) emit("cancelled", false);
-      throw error;
-    }
-  }
-
-  pendingMutation(): PendingMutationSummary | null {
-    const pending = parseStored<PendingMutation>(this.storage.getItem(this.pendingMutationKey()));
-    const token = this.currentToken();
-    if (!pending || !token
-        || pending.collectionId !== token.collectionId
-        || pending.grantId !== token.grantId
-        || pending.keyId !== token.encryption?.key_id) return null;
-    return { operation: pending.operation, createdAt: pending.createdAt, resumable: true };
-  }
-
-  async resumePendingMutation<Result>(
-    input: unknown,
-    options?: OperationRequestOptions
-  ): Promise<Result> {
-    const pending = this.pendingMutation();
-    if (!pending) {
-      throw new MdbaseConnectError(
-        "no_pending_mutation",
-        "There is no interrupted mutation to resume.",
-        { recovery: "none" }
-      );
-    }
-    return this.performOperation<Result>(pending.operation, input, options);
-  }
-
-  validate(input: JsonObject = {}): Promise<MdbaseOperationEnvelope> {
-    return this.collectionClient.validate(input);
-  }
-
-  readType(input: ReadTypeInput): Promise<MdbaseOperationEnvelope<CollectionTypeDocument>> {
-    return this.collectionClient.readType(input);
-  }
-
-  createType(input: CreateTypeInput): Promise<MdbaseOperationEnvelope<CollectionTypeDocument>> {
-    return this.collectionClient.createType(input);
-  }
-
-  updateType(input: UpdateTypeInput): Promise<MdbaseOperationEnvelope<CollectionTypeDocument>> {
-    return this.collectionClient.updateType(input);
-  }
-
-  installTypePack(input: TypePackProvision): Promise<MdbaseOperationEnvelope<TypePackInstallResult>> {
-    return this.collectionClient.installTypePack(input);
-  }
-
-  listTimers(namespace: string): Promise<MdbaseTimerList> {
-    return this.collectionClient.listTimers(namespace);
-  }
-
-  putTimer(input: {
-    namespace: string;
-    criterion_id: string;
-    timer: MdbaseDesiredTimer;
-  }): Promise<MdbaseTimer> {
-    return this.collectionClient.putTimer(input);
-  }
-
-  cancelTimer(input: {
-    namespace: string;
-    id: string;
-    generation?: number;
-  }): Promise<{ namespace: string; id: string; cancelled: boolean }> {
-    return this.collectionClient.cancelTimer(input);
-  }
-
-  reconcileTimers(input: {
-    namespace: string;
-    criterion_id: string;
-    timers: MdbaseDesiredTimer[];
-  }): Promise<MdbaseTimerReconciliation> {
-    return this.collectionClient.reconcileTimers(input);
-  }
-
-  async *watch(options: WatchOptions = {}): AsyncGenerator<CollectionChange> {
-    yield* this.collectionClient.watch(options);
-  }
-
-  async operation<Result>(
-    operation: CollectionOperation,
-    input: unknown,
-    options?: OperationRequestOptions
-  ): Promise<Result> {
-    return this.collectionClient.operation(operation, input, options);
-  }
-
-  private async performOperation<Result>(
-    operation: CollectionOperation,
-    input: unknown,
-    options: OperationRequestOptions = {}
-  ): Promise<Result> {
-    throwIfCancelled(options.signal);
-    let token = this.currentToken();
-    if (!token) throw new MdbaseConnectError("not_authorized", "Connect this application before accessing a collection.");
-    if (!token.operations.includes(operation)) {
-      throw new MdbaseConnectError("insufficient_access", `This connection does not allow ${operation}.`, {
-        requiresAuthorization: true,
-        recovery: "reauthorize",
-        details: {
-          requiredOperations: [operation],
-          grantedOperations: [...token.operations],
-          missingOperations: [operation]
-        }
-      });
-    }
-    let tryDirect = await this.shouldAttemptDirect(token);
-    if (!tryDirect) {
-      token = await this.authorizedToken();
-      if (!token) throw new MdbaseConnectError("not_authorized", "Reconnect this application to continue.");
-    }
-    let attempt: OperationAttempt;
-    try {
-      attempt = await this.sendOperation(token, operation, input, tryDirect, options);
-    } catch (error) {
-      throw operationTransportError(error, options.signal, isMutation(operation, input) && this.pendingMutation() !== null);
-    }
-    let response = attempt.response;
-    const staleBinding = response.status === 409
-      && (await response.clone().json().catch(() => null))?.error?.code === "encryption_binding_stale";
-    if ((response.status === 401 || staleBinding) && token.refreshToken) {
-      if (attempt.pendingMutation
-          && (attempt.directDeliveryUncertain
-            || (attempt.encryptedRequest && attempt.resumingMutation))) {
-        throw new MdbaseConnectError(
-          "direct_outcome_unknown",
-          "The direct operation may have completed, but its encrypted grant changed before the response could be recovered. Refresh before making another change."
-        );
-      }
-      if (attempt.pendingMutation) this.clearPendingMutation();
-      token = await this.refreshAuthorization();
-      tryDirect = await this.shouldAttemptDirect(token);
-      try {
-        attempt = await this.sendOperation(token, operation, input, tryDirect, options);
-      } catch (error) {
-        throw operationTransportError(error, options.signal, isMutation(operation, input) && this.pendingMutation() !== null);
-      }
-      response = attempt.response;
-    }
-    const body = await response.json();
-    if (!response.ok) {
-      const error = apiError(body, "operation_failed", "Collection operation failed.", response.status);
-      if (attempt.pendingMutation
-          && (attempt.directDeliveryUncertain
-            || (attempt.encryptedRequest && attempt.resumingMutation))) {
-        throw uncertainDirectMutation(error);
-      }
-      if (attempt.pendingMutation && !attempt.directDeliveryUncertain) {
-        this.clearPendingMutation();
-      }
-      if (error.code === "direct_operation_rejected" && error.status === 403) {
-        this.invalidateRejectedAuthorization(token);
-      }
-      throw error;
-    }
-    if (attempt.encryptedRequest) {
-      const encryptedResponse = body?.envelope as EncryptedRelayOperationResponse | undefined;
-      if (!encryptedResponse || !token.encryption || !token.grantId || !token.keyHandle) {
-        throw new MdbaseConnectError(
-          "invalid_encrypted_response",
-          "The relay did not return an encrypted connector response."
-        );
-      }
-      try {
-        const decrypted = await decryptRelayResponse<Result>(
-          this.keyStore,
-          token.keyHandle,
-          { grantId: token.grantId, applicationId: token.clientId, encryption: token.encryption },
-          attempt.encryptedRequest,
-          encryptedResponse
-        );
-        if (attempt.pendingMutation) this.clearPendingMutation();
-        if (!decrypted.ok) throw new MdbaseConnectError(decrypted.error.code, decrypted.error.message);
-        return decrypted.result;
-      } catch (error) {
-        if (error instanceof MdbaseConnectError) throw error;
-        if (error instanceof RelayCryptoError) throw new MdbaseConnectError(error.code, error.message);
-        throw error;
-      }
-    }
-    if (body?.protocol_version !== 1 || body?.request_id !== attempt.requestId) {
-      throw new MdbaseConnectError(
-        "invalid_operation_response",
-        "The collection authority returned a response for a different protocol request."
-      );
-    }
-    if (attempt.pendingMutation) this.clearPendingMutation();
-    return body.result as Result;
-  }
-
-  private async performAuthoritySyncRequest<Result>(
-    collectionId: string,
-    replicaId: string,
-    method: "GET" | "POST",
-    path: string,
-    input?: unknown
-  ): Promise<Result> {
-    let token = await this.authorizedToken();
-    if (!token?.authority
-        || token.collectionId !== collectionId
-        || token.authority.replicaId !== replicaId) {
-      throw new MdbaseConnectError(
-        "authority_authorization_changed",
-        "Reconnect this collection authority before synchronizing."
-      );
-    }
-    let response = await this.sendAuthoritySyncRequest(token, method, path, input);
-    if (response.status === 401 && token.refreshToken) {
-      token = await this.refreshAuthorization();
-      if (!token.authority
-          || token.collectionId !== collectionId
-          || token.authority.replicaId !== replicaId) {
-        throw new MdbaseConnectError(
-          "authority_authorization_changed",
-          "Reconnect this collection authority before synchronizing."
-        );
-      }
-      response = await this.sendAuthoritySyncRequest(token, method, path, input);
-    }
-    const body = await response.json();
-    if (!response.ok) throw apiError(body, "sync_failed", "Collection synchronization failed.", response.status);
-    return body as Result;
-  }
-
-  private async sendAuthoritySyncRequest(
-    token: StoredToken,
-    method: "GET" | "POST",
-    path: string,
-    input?: unknown
-  ): Promise<Response> {
-    if (!token.authority) {
-      throw new MdbaseConnectError("not_remote_authority", "This authorization has no remote authority endpoint.");
-    }
-    const url = `${token.authority.syncUrl}/${path}`;
-    const body = input === undefined ? undefined : JSON.stringify(input);
-    const proof = await this.authorityProofHeaders(token, method, url, body, token.authority.accessToken);
-    return fetch(
-      url,
-      {
-        method,
-        headers: {
-          authorization: `Bearer ${token.authority.accessToken}`,
-          ...(input === undefined ? {} : { "content-type": "application/json" }),
-          ...proof
-        },
-        ...(body === undefined ? {} : { body })
-      }
-    );
-  }
-
-  private async sendOperation(
-    token: StoredToken,
-    operation: CollectionOperation,
-    input: unknown,
-    tryDirect: boolean,
-    options: OperationRequestOptions = {}
-  ): Promise<OperationAttempt> {
-    let body: unknown = input ?? {};
-    let encryptedRequest: Awaited<ReturnType<typeof encryptRelayRequest>> | undefined;
-    let pendingMutation = false;
-    let resumingMutation = false;
-    let requestId: string = crypto.randomUUID();
-    let pending: PendingMutation | null = null;
-    let inputFingerprint: string | undefined;
-    if (isMutation(operation, input)) {
-      inputFingerprint = await operationFingerprint(operation, input);
-      pending = parseStored<PendingMutation>(
-        this.storage.getItem(this.pendingMutationKey())
-      );
-      if (pending) {
-        if (pending.collectionId !== token.collectionId
-            || pending.grantId !== token.grantId
-            || pending.keyId !== token.encryption?.key_id
-            || pending.operation !== operation
-            || pending.inputFingerprint !== inputFingerprint) {
-          throw new MdbaseConnectError(
-            "pending_mutation_unresolved",
-            "A previous write still has an unknown outcome. Retry that exact write before making another change."
-          );
-        }
-        requestId = pending.requestId;
-        resumingMutation = true;
-      }
-      pendingMutation = true;
-    }
-    if (token.encryption && !token.authority) {
-      if (!token.grantId || !token.keyHandle) {
-        throw new MdbaseConnectError("missing_grant_key", "Reconnect this application to restore encrypted access.");
-      }
-      try {
-        if (pendingMutation) {
-          if (pending) {
-            if (!pending.envelope) {
-              throw new MdbaseConnectError(
-                "pending_mutation_unresolved",
-                "The pending write belongs to a different transport. Reconnect before retrying it."
-              );
-            }
-            encryptedRequest = pending.envelope;
-          } else {
-            encryptedRequest = await encryptRelayRequest(
-              this.keyStore,
-              token.keyHandle,
-              { grantId: token.grantId, applicationId: token.clientId, encryption: token.encryption },
-              operation,
-              input,
-              requestId
-            );
-            this.storage.setItem(this.pendingMutationKey(), JSON.stringify({
-              collectionId: token.collectionId,
-              ...(token.grantId ? { grantId: token.grantId } : {}),
-              keyId: token.encryption.key_id,
-              operation,
-              inputFingerprint: inputFingerprint!,
-              requestId,
-              envelope: encryptedRequest,
-              createdAt: Date.now()
-            } satisfies PendingMutation));
-          }
-        } else {
-          encryptedRequest = await encryptRelayRequest(
-            this.keyStore,
-            token.keyHandle,
-            { grantId: token.grantId, applicationId: token.clientId, encryption: token.encryption },
-            operation,
-            input
-          );
-        }
-      } catch (error) {
-        if (error instanceof RelayCryptoError) throw new MdbaseConnectError(error.code, error.message);
-        throw error;
-      }
-      body = encryptedRequest;
-    } else {
-      const request: MdbaseOperationRequest = {
-        protocol_version: 1,
-        request_id: requestId,
-        input: body
-      };
-      body = request;
-      if (pendingMutation && !pending) {
-        this.storage.setItem(this.pendingMutationKey(), JSON.stringify({
-          collectionId: token.collectionId,
-          ...(token.grantId ? { grantId: token.grantId } : {}),
-          operation,
-          inputFingerprint: inputFingerprint!,
-          requestId,
-          createdAt: Date.now()
-        } satisfies PendingMutation));
-      }
-    }
-    if (tryDirect && encryptedRequest && !token.authority) {
-      let directDeliveryUncertain = false;
-      try {
-        const response = await fetch(`${this.loopbackUrl}/v1/operations`, loopbackRequest({
-          method: "POST",
-          headers: { "content-type": "application/mdbase-connect+json" },
-          body: JSON.stringify(encryptedRequest),
-          signal: options.signal
-        }));
-        if (!directFallbackStatus(response.status)) {
-          if (response.ok) {
-            this.markDirectAvailable();
-            this.setRoute("direct");
-          }
-          return {
-            response,
-            requestId,
-            encryptedRequest,
-            pendingMutation,
-            resumingMutation
-          };
-        }
-        directDeliveryUncertain = response.status >= 500;
-        this.markDirectUnavailable();
-      } catch (error) {
-        if (options.signal?.aborted) throw error;
-        directDeliveryUncertain = true;
-        if ((await localNetworkPermission()) === "denied") this.setDirectStatus("denied");
-        else this.markDirectUnavailable();
-      }
-      let relayToken: StoredToken;
-      try {
-        relayToken = token.expiresAt > Date.now() + 30_000
-          ? token
-          : await this.refreshAuthorization();
-      } catch (error) {
-        if (pendingMutation && (directDeliveryUncertain || resumingMutation)) throw uncertainDirectMutation(error);
-        throw error;
-      }
-      let response: Response;
-      try {
-        response = await fetch(
-          `${this.serverUrl}/v1/authorities/${encodeURIComponent(relayToken.collectionId)}/operations/${operation}`,
-          {
-          method: "POST",
-          headers: {
-            authorization: `Bearer ${relayToken.accessToken}`,
-            "content-type": "application/json"
-          },
-          body: JSON.stringify(encryptedRequest),
-          signal: options.signal
-          }
-        );
-      } catch (error) {
-        if (pendingMutation && (directDeliveryUncertain || resumingMutation)) throw uncertainDirectMutation(error);
-        throw error;
-      }
-      if (response.ok) this.setRoute("relay");
-      return {
-        response,
-        requestId,
-        encryptedRequest,
-        directDeliveryUncertain,
-        pendingMutation,
-        resumingMutation
-      };
-    }
-    const operationUrl = token.authority
-      ? `${token.authority.operationsUrl}/${operation}`
-      : `${this.serverUrl}/v1/authorities/${encodeURIComponent(token.collectionId)}/operations/${operation}`;
-    const operationBody = JSON.stringify(body);
-    const proof = token.authority
-      ? await this.authorityProofHeaders(
-          token,
-          "POST",
-          operationUrl,
-          operationBody,
-          token.authority.accessToken
-        )
-      : {};
-    const response = await fetch(operationUrl, {
-        method: "POST",
-        headers: {
-          authorization: `Bearer ${token.authority?.accessToken ?? token.accessToken}`,
-          "content-type": "application/json",
-          ...proof
-        },
-        body: operationBody,
-        signal: options.signal
-      });
-    if (response.ok) this.setRoute(token.authority ? "remote" : "relay");
-    return { response, requestId, encryptedRequest, pendingMutation, resumingMutation };
-  }
-
-  private directCapable(token: StoredToken | null): boolean {
-    if (!token || token.authority || !token.encryption || !token.grantId || !token.keyHandle) return false;
-    if (this.directAccessMode === "disabled") return false;
-    if (typeof location !== "undefined"
-        && token.applicationOrigin
-        && token.applicationOrigin !== location.origin) return false;
-    return true;
-  }
-
-  private hasResumableMutationTransport(): boolean {
-    const token = this.currentToken();
-    return Boolean(token);
-  }
-
-  private directEligible(token: StoredToken | null): token is StoredToken {
-    return token !== null
-      && this.directCapable(token)
-      && this.storage.getItem(this.directPreferenceKey()) !== "disabled";
-  }
-
-  private async shouldAttemptDirect(token: StoredToken): Promise<boolean> {
-    if (!this.directEligible(token)) return false;
-    if (this.directStatus === "available") return true;
-    if (this.directStatus === "unavailable" && Date.now() < this.directRetryAt) return false;
-    const permission = await localNetworkPermission();
-    if (permission === "denied") {
-      this.setDirectStatus("denied");
-      return false;
-    }
-    if ((permission === "prompt" || permission === null)
-        && this.storage.getItem(this.directPreferenceKey()) !== "enabled") {
-      this.setDirectStatus("permission_required");
-      return false;
-    }
-    this.setDirectStatus("checking");
-    return true;
-  }
-
-  private async probeDirectAccess(): Promise<DirectAccessStatus> {
-    this.setDirectStatus("checking");
-    try {
-      const response = await fetch(`${this.loopbackUrl}/v1/ready`, loopbackRequest({
-        method: "GET",
-        cache: "no-store"
-      }));
-      const body = await response.json().catch(() => null);
-      if (response.ok
-          && body?.service === "mdbase-connect"
-          && body?.loopback_protocol_version === 1
-          && body?.encrypted_protocol_version === 1) {
-        this.markDirectAvailable();
-        return "available";
-      }
-    } catch {
-      // Permission denial, unsupported mixed content, and an absent connector all reject fetch.
-    }
-    if ((await localNetworkPermission()) === "denied") return this.setDirectStatus("denied");
-    this.markDirectUnavailable();
-    return "unavailable";
-  }
-
-  private markDirectAvailable(): void {
-    this.directFailures = 0;
-    this.directRetryAt = 0;
-    this.setDirectStatus("available");
-  }
-
-  private markDirectUnavailable(): void {
-    this.directFailures += 1;
-    this.directRetryAt = Date.now() + Math.min(60_000, 1_000 * 2 ** (this.directFailures - 1));
-    this.setDirectStatus("unavailable");
-  }
-
-  private setDirectStatus(status: DirectAccessStatus): DirectAccessStatus {
-    if (this.directStatus !== status) {
-      this.directStatus = status;
-      this.emitConnection();
-    }
-    return status;
-  }
-
-  private setRoute(route: MdbaseConnectionRoute): void {
-    if (this.currentRoute !== route) {
-      this.currentRoute = route;
-      this.emitConnection();
-    }
-  }
-
-  private emitConnection(): void {
-    const connection = this.info();
-    for (const listener of this.connectionListeners) listener(connection);
-  }
-
-  private invalidateRejectedAuthorization(rejected: StoredToken): void {
-    const current = parseStored<StoredToken>(this.storage.getItem(this.tokenKey()));
-    if (!current || !sameAuthorization(current, rejected)) return;
-    this.internals.removeToken(this.collectionId, current.keyHandle);
-    this.currentRoute = "relay";
-    this.directStatus = this.directAccessMode === "disabled" ? "disabled" : "unavailable";
-    this.emitConnection();
-  }
-
-  private currentToken(): StoredToken | null {
-    const stored = this.storage.getItem(this.tokenKey());
-    const token = parseStored<StoredToken>(stored);
-    const invalidate = (keyHandle?: unknown): null => {
-      this.internals.removeToken(
-        this.collectionId,
-        typeof keyHandle === "string" ? keyHandle : undefined,
-        "invalid_stored_grant"
-      );
-      return null;
-    };
-    if (!token) {
-      if (stored) invalidate();
-      return null;
-    }
-    if (
-      token.version !== 1
-      || typeof token.accessToken !== "string"
-      || token.accessToken.length === 0
-      || typeof token.clientId !== "string"
-      || token.clientId.length === 0
-      || token.collectionId !== this.collectionId
-      || typeof token.collectionName !== "string"
-      || token.collectionName.length === 0
-      || !Array.isArray(token.operations)
-      || token.operations.some((operation) => typeof operation !== "string")
-      || typeof token.expiresAt !== "number"
-      || !Number.isFinite(token.expiresAt)
-      || (
-        token.refreshToken !== undefined
-        && (
-          typeof token.refreshToken !== "string"
-          || token.refreshToken.length === 0
-        )
-      )
-      || (
-        token.refreshExpiresAt !== undefined
-        && (
-          typeof token.refreshExpiresAt !== "number"
-          || !Number.isFinite(token.refreshExpiresAt)
-        )
-      )
-    ) return invalidate(token.keyHandle);
-    if (!parseGrantScope(token.scope)) {
-      return invalidate(token.keyHandle);
-    }
-    if (
-      token.authority
-      && !validStoredAuthority(token.authority, token.collectionId)
-    ) {
-      return invalidate(token.keyHandle);
-    }
-    if (this.internals.relayEncryption === "required") {
-      if (token.authority) {
-        if (!token.keyHandle || !token.authority.proofPublicKey) {
-          return invalidate(token.keyHandle);
-        }
-      } else {
-        if (
-          !token.grantId
-          || !token.keyHandle
-          || !validStoredEncryption(token.encryption, token.collectionId)
-        ) {
-          return invalidate(token.keyHandle);
-        }
-      }
-    }
-    if (token.expiresAt <= Date.now()
-        && (!token.refreshToken || (token.refreshExpiresAt ?? 0) <= Date.now())) {
-      // The cloud bearer and the local grant proof have separate lifetimes. Keep an
-      // encrypted local grant usable while the connector still recognizes it; relay
-      // use will require reauthorization, and revocation remains enforced locally.
-      if (this.directCapable(token)) return token;
-      this.internals.removeToken(this.collectionId, token.keyHandle);
-      return null;
-    }
-    return token;
-  }
-
-  private async authorizedToken(): Promise<StoredToken | null> {
-    const token = this.currentToken();
-    if (!token) return null;
-    if (token.expiresAt > Date.now() + 30_000) return token;
-    if (!token.refreshToken || (token.refreshExpiresAt ?? 0) <= Date.now()) {
-      if (this.directCapable(token)) {
-        throw new MdbaseConnectError(
-          "relay_authorization_expired",
-          "Direct access is still available on this computer, but using the relay requires reconnecting this application."
-        );
-      }
-      this.internals.removeToken(this.collectionId, token.keyHandle);
-      return null;
-    }
-    return this.refreshAuthorization();
-  }
-
-  private refreshAuthorization(): Promise<StoredToken> {
-    if (this.refreshPromise) return this.refreshPromise;
-    this.refreshPromise = this.performRefresh().finally(() => {
-      this.refreshPromise = null;
-    });
-    return this.refreshPromise;
-  }
-
-  private async performRefresh(): Promise<StoredToken> {
-    const current = this.currentToken();
-    if (!current?.refreshToken) {
-      throw new MdbaseConnectError("not_authorized", "Reconnect this application to continue.");
-    }
-    if ((current.refreshExpiresAt ?? 0) <= Date.now()) {
-      throw new MdbaseConnectError(
-        "relay_authorization_expired",
-        "Direct access is still available on this computer, but using the relay requires reconnecting this application."
-      );
-    }
-    const attemptedRefreshToken = current.refreshToken;
-    const refreshUrl = `${this.serverUrl}/oauth/token`;
-    const refreshBody = new URLSearchParams({
-      grant_type: "refresh_token",
-      refresh_token: current.refreshToken,
-      client_id: current.clientId
-    }).toString();
-    const proof = current.authority
-      ? await this.authorityProofHeaders(
-          current,
-          "POST",
-          refreshUrl,
-          refreshBody,
-          current.refreshToken
-        )
-      : {};
-    const response = await fetch(refreshUrl, {
-      method: "POST",
-      headers: {
-        "content-type": "application/x-www-form-urlencoded",
-        ...proof
-      },
-      body: refreshBody
-    });
-    const body = await response.json();
-    if (!response.ok) {
-      const latest = this.currentToken();
-      if (latest?.refreshToken && latest.refreshToken !== attemptedRefreshToken) {
-        return latest;
-      }
-      if (!this.directCapable(current)) {
-        this.internals.removeToken(this.collectionId, current.keyHandle);
-      }
-      throw apiError(body, "authorization_expired", "Reconnect this application to continue.", response.status);
-    }
-    return this.storeTokenResponse(body, current.clientId, current.keyHandle);
-  }
-
-  private async authorityProofHeaders(
-    token: StoredToken,
-    method: string,
-    url: string,
-    body: string | undefined,
-    credential: string
-  ): Promise<Record<string, string>> {
-    if (!token.authority?.proofPublicKey) return {};
-    if (!token.keyHandle) {
-      throw new MdbaseConnectError(
-        "missing_grant_key",
-        "Reconnect this application to restore remote authority request signing."
-      );
-    }
-    try {
-      const target = new URL(url);
-      return await signAuthorityRequest(
-        this.keyStore,
-        token.keyHandle,
-        token.authority.proofPublicKey,
-        {
-          method,
-          target: `${target.pathname}${target.search}`,
-          body,
-          credential
-        }
-      );
-    } catch (error) {
-      if (error instanceof RelayCryptoError) {
-        throw new MdbaseConnectError(error.code, error.message);
-      }
-      throw error;
-    }
-  }
-
-  private storeTokenResponse(body: any, clientId: string, keyHandle?: string): StoredToken {
-    if (body.collection_id !== this.collectionId) {
-      throw new MdbaseConnectError(
-        "collection_mismatch",
-        "The refreshed authorization belongs to a different collection."
-      );
-    }
-    const token = this.internals.storeTokenResponse(body, clientId, keyHandle);
-    this.currentRoute = token.authority ? "remote" : "relay";
-    this.directStatus = token.authority || this.directAccessMode === "disabled"
-      ? "disabled"
-      : "unavailable";
-    this.emitConnection();
-    return token;
-  }
-
-  private tokenKey() {
-    return this.internals.tokenKey(this.collectionId);
-  }
-  private notificationKey(transport: "web_push" | "fcm" = "web_push") {
-    return this.internals.notificationKey(this.collectionId, transport);
-  }
-  private async deleteNotificationChannel(
-    channelId: string,
-    accessToken: string
-  ): Promise<void> {
-    const response = await fetch(
-      `${this.serverUrl}/v1/notifications/channels/${channelId}`,
-      {
-        method: "DELETE",
-        headers: { authorization: `Bearer ${accessToken}` }
-      }
-    );
-    if (!response.ok && response.status !== 404) {
-      const body = await response.json();
-      throw apiError(
-        body,
-        "notification_unregistration_failed",
-        "Could not unregister push notifications.",
-        response.status
-      );
-    }
-  }
-  private pendingMutationKey() {
-    return this.internals.pendingMutationKey(this.collectionId);
-  }
-  private clearPendingMutation(): void {
-    this.storage.removeItem(this.pendingMutationKey());
-  }
-  private directPreferenceKey() {
-    return this.internals.directPreferenceKey();
-  }
-}
-
-function sessionSnapshotKey<Frontmatter extends JsonObject>(
-  snapshot: MdbaseSessionSnapshot<Frontmatter>
-): string {
-  if (snapshot.status === "unselected") {
-    return JSON.stringify({ status: snapshot.status, connections: snapshot.connections });
-  }
-  if (snapshot.status === "unavailable") {
-    return JSON.stringify({
-      status: snapshot.status,
-      collectionId: snapshot.collectionId,
-      reason: snapshot.reason,
-      connections: snapshot.connections
-    });
-  }
-  return JSON.stringify({
-    status: snapshot.status,
-    collectionId: snapshot.collectionId,
-    info: snapshot.info,
-    access: snapshot.access,
-    connections: snapshot.connections
-  });
-}
